+++ src/com/intellij/idea/plugin/hybris/diagram/module/node/graph/ModuleDepGraphFactory.kt
@@ -24,29 +24,34 @@ import com.intellij.openapi.module.Module
 
 object ModuleDepGraphFactory {
 
-    fun buildNode(module: Module): ModuleDepGraphNodeModule {
+    fun buildNode(module: Module) = with(HybrisProjectSettingsComponent.getInstance(module.project).getModuleSettings(module)) {
         val properties = mutableListOf<ModuleDepGraphField>()
 
-        val moduleSettings = HybrisProjectSettingsComponent.getInstance(module.project).getModuleSettings(module)
+        if (description != null) properties.add(ModuleDepGraphFieldDescription(description!!))
+        if (version != null) properties.add(ModuleDepGraphFieldParameter("Version", version))
 
-        if (moduleSettings.description != null) properties.add(ModuleDepGraphFieldDescription(moduleSettings.description!!))
-        if (moduleSettings.subModuleType != null) properties.add(ModuleDepGraphFieldSubModuleType("Sub-module Type", moduleSettings.subModuleType!!.name))
-        if (moduleSettings.deprecated) properties.add(ModuleDepGraphFieldParameter("Deprecated"))
-        if (moduleSettings.useMaven) properties.add(ModuleDepGraphFieldParameter("Maven Enabled"))
-        if (moduleSettings.jaloLogicFree) properties.add(ModuleDepGraphFieldParameter("Jalo Logic Free"))
-        if (moduleSettings.extGenTemplateExtension) properties.add(ModuleDepGraphFieldParameter("Template Extension"))
-        if (moduleSettings.moduleGenName != null) properties.add(ModuleDepGraphFieldParameter("Module Generation Name"))
-        if (moduleSettings.hacModule) properties.add(ModuleDepGraphFieldParameter("HAC module"))
-        if (moduleSettings.webModule) properties.add(ModuleDepGraphFieldParameter("Web module"))
-        if (moduleSettings.hmcModule) properties.add(ModuleDepGraphFieldParameter("HMC module"))
-        if (moduleSettings.backofficeModule) properties.add(ModuleDepGraphFieldParameter("Backoffice module"))
-        if (moduleSettings.addon) properties.add(ModuleDepGraphFieldParameter("Addon"))
-        if (moduleSettings.classPathGen != null) properties.add(ModuleDepGraphFieldParameter("Classpath Generation"))
+        if (deprecated) properties.add(ModuleDepGraphFieldParameter("Deprecated"))
+        if (useMaven) properties.add(ModuleDepGraphFieldParameter("Maven Enabled"))
+        if (jaloLogicFree) properties.add(ModuleDepGraphFieldParameter("Jalo Logic Free"))
+        if (extGenTemplateExtension) properties.add(ModuleDepGraphFieldParameter("Template Extension"))
+        if (requiredByAll) properties.add(ModuleDepGraphFieldParameter("Required by All"))
 
-        return ModuleDepGraphNodeModule(
+        if (subModuleType != null) properties.add(ModuleDepGraphFieldParameter("Sub-module Type", subModuleType!!.name))
+        if (moduleGenName != null) properties.add(ModuleDepGraphFieldParameter("Module Generation Name", moduleGenName))
+        if (classPathGen != null) properties.add(ModuleDepGraphFieldParameter("Classpath Generation", classPathGen))
+
+        if (coreModule) properties.add(ModuleDepGraphFieldParameter("Core module", packageRoot))
+        if (webModule) properties.add(ModuleDepGraphFieldParameter("Web module", webRoot))
+
+        if (backofficeModule) properties.add(ModuleDepGraphFieldParameter("Backoffice module"))
+        if (hacModule) properties.add(ModuleDepGraphFieldParameter("HAC module"))
+        if (hmcModule) properties.add(ModuleDepGraphFieldParameter("HMC module"))
+        if (addon) properties.add(ModuleDepGraphFieldParameter("Addon"))
+
+        ModuleDepGraphNodeModule(
             module,
-            moduleSettings.type,
-            moduleSettings.subModuleType,
+            type,
+            subModuleType,
             module.yExtensionName(),
             properties.toTypedArray()
         )