+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Added possibility to display extension details within the Module node [#1041](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1041)
 - Added nodes colorization based on the extension type [#1042](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1042)
 - Added edges colorization based on the extension type [#1043](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1043)
+- Show additional extension details within the Module node [#1044](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1044)
 
 ### `Groovy` enhancements
 - Actions toolbar for disabled for Groovy IDE Scripts by default [#1040](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1040)